requirements.txt
appdirs==1.4.4
asgiref==3.5.2
autopep8==1.6.0

certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Deprecated==1.2.13
distlib==0.3.1
Django==3.2
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-crontab==0.7.1
django-db-logger==0.1.10
django-environ==0.9.0
django-filter==21.1
django-jsonfield==1.4.1
django-mathfilters==1.0.0
django-redis==5.2.0
django-rq==2.5.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-user-agents
et-xmlfile==1.1.0
filelock==3.0.12
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
install==1.3.5
isodate==0.6.1
jsonfield==3.1.0
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
numpy==1.22.2
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.2.0
plotly==5.6.0
portalocker==2.3.2
psycopg2-binary
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
PyJWT==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
redis==4.3.3
reportlab==3.6.11
requests==2.27.1
requests-oauthlib==1.3.1
rq==1.11.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
toml==0.10.2
typing_extensions==4.2.0
tzdata==2022.2
urllib3==1.26.8
virtualenv==20.4.2
wrapt==1.14.1
XlsxWriter==3.0.3
django-user-sessions
wheel
bump2version
twine
wizard
django_otp>=0.8.0
qrcode>=4.0.0,<7.99
django-phonenumber-field>=1.1.0,<7
django-formtools
twilio
phonenumberslite
django-otp-yubikey
webauthn>=1.6.0,<1.99
